requirements.txt
selenium==4.38.0
webdriver-manager==4.0.2
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.2.1
pywin32==311
